--- backend/src/lib/aws/index.ts
@@ -1,1 +1,0 @@
-export * from "./elasticache";
